Fix native_lab result handling and add new runtime results documentation
- Updated `native_lab_runner` to correctly read results from skb context for TC and cgroup programs, ensuring compatibility with kernel BPF semantics.
- Introduced a new README file documenting the results of the successful full native/kernel/native_lab run, including performance comparisons across different execution modes.
- Clarified the root cause of previous failures and provided a summary of the performance metrics for the tested micro programs.

+++ micro/programs/payload_prefix_memcmp_scan.bpf.c
@@ -4,9 +4,9 @@
 #define MEMCMP_PREFIX_LEN 64U
 #define MEMCMP_PREFIX_INPUT_SIZE (4U + MEMCMP_PREFIX_SCENARIOS * MEMCMP_PREFIX_LEN)
 
-static __always_inline u8 memcmp_prefix_pattern_byte(u32 index)
+static __always_inline u32 memcmp_prefix_pattern_byte(u32 index)
 {
-    return (u8)((((index * 29U) ^ (index << 2U) ^ 0xA5U) + 0x11U) & 0xFFU);
+    return (((index * 29U) ^ (index << 2U) ^ 0xA5U) + 0x11U) & 0xFFU;
 }
 
 static __always_inline int bench_payload_prefix_memcmp_scan(const u8 *data, u32 len, u64 *out)
